@@ -14,7 +14,7 @@
     "@types/figures": "^2.0.0",
     "@types/inquirer": "^0.0.35",
     "@types/node": "^8.0.47",
-    "typescript": "^2.5.3"
+    "typescript": "^4.5.5"
   },
   "dependencies": {
     "chalk": "^2.3.0",
